# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Fidelito API
  version: 2.0.0
  description: |
    Digital loyalty card platform for Tunisian businesses.
    Clients display dynamic TOTP QR codes; cashiers scan to stamp loyalty cards.
    Businesses manage staff, rewards, and analytics. Admins manage merchants.
servers:
  - url: http://localhost:8787
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/business/staff:
    get:
      summary: List staff members
      responses:
        '200': { description: Staff list }
    post:
      summary: Add staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
                role: { type: string, enum: [owner, cashier], default: cashier }
      responses:
        '200': { description: Added }
        '404': { description: User not found }
        '409': { description: Already staff }
  /api/business/staff/{userId}:
    delete:
      summary: Remove staff member
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Removed }
    patch:
      summary: Update staff role
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [owner, cashier] }
      responses:
        '200': { description: Updated }
...
  /api/business/transactions:
    get:
      summary: All transactions (paginated)
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        '200': { description: Transactions }
  /api/business/rewards:
    get:
      summary: List rewards
      responses:
        '200': { description: Rewards }
    post:
      summary: Create reward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                stampsCost: { type: integer, default: 10 }
      responses:
        '201': { description: Created }
  /api/business/rewards/{id}:
    patch:
      summary: Update reward
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                stampsCost: { type: integer }
                isAvailable: { type: boolean }
      responses:
        '200': { description: Updated }
    delete:
      summary: Delete reward
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Deleted }
  /api/business/settings:
    patch:
      summary: Update merchant settings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                stampsPerReward: { type: integer }
                monthlyPointCap: { type: integer }
                logoUrl: { type: string, nullable: true }
      responses:
        '200': { description: Updated }

  # --- Profile ---
...